% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Detection_P_Ship_Based_Surveys.R
\name{Calculate_Detection_P_Ship_Based_Surveys}
\alias{Calculate_Detection_P_Ship_Based_Surveys}
\title{Execute distance analysis on ship-based survey results}
\usage{
Calculate_Detection_P_Ship_Based_Surveys(
  esas_table_2_analyse,
  species_2_analyse
)
}
\arguments{
\item{esas_table_2_analyse}{A data frame containing survey data with columns
for distance bins, platform class, transect type, observation distance,
species code, behavior, and count. As returned by \code{\link[=Create_ESAS_Table]{Create_ESAS_Table()}}.}

\item{species_2_analyse}{A vector of species codes as encoded in the species
column of the Observations table in the ESAS Data Model. See the
\href{https://esas-docs.ices.dk/species/}{Species} page of the Data Model.}
}
\value{
A data.frame with the following columns:
\itemize{
\item Species: The species code.
\item Function: The selected detection function ("HR" for hazard-rate or "HN" for half-normal).
\item Detection_P_AVG: The average detection probability for the species.
}
}
\description{
This function calculates detection probabilities for specified species from
ship-based survey data using distance sampling methods. It filters the input
data for relevant observations, fits half-normal and hazard-rate models, and
selects the best model based on AIC values. The output is a data frame
containing species names, the selected model function, and the average
detection probability.
}
\examples{
# Read 4 ESAS tables:
path_to_read <- system.file("extdata", "ESAS_0827343782", package = "esas")
esas_tables_list <- Read_ESAS_Tables(
  path = path_to_read,
  file_encoding = "UTF-8"
)

# Create an ESAS master-table:
esas_table <- Create_ESAS_Table(esas_tables_list = esas_tables_list)


# Execute distance analysis on selection of species:
Calculate_Detection_P_Ship_Based_Surveys(
  esas_table_2_analyse = esas_table,
  species_2_analyse = c(720, 6020)
)
}
\seealso{
Other analysis functions: 
\code{\link{Create_ESAS_Table}()},
\code{\link{Create_Seabird_Density_Cross_Table}()}
}
\concept{analysis functions}
