% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esas-package.R
\docType{package}
\name{esas-package}
\alias{esas}
\alias{esas-package}
\title{esas: Read and Process Data from European Seabirds At Sea}
\description{
Read and process data from European Seabirds At Sea ('ESAS').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/esas}
  \item \url{https://inbo.github.io/esas}
  \item Report bugs at \url{https://github.com/inbo/esas/issues}
}

}
\author{
\strong{Maintainer}: Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID}) (Research Institute for Nature and Forest (INBO))

Authors:
\itemize{
  \item Nicolas Vanermen \email{nicolas.vanermen@inbo.be} (\href{https://orcid.org/0000-0003-0156-357X}{ORCID}) (Research Institute for Nature and Forest (INBO))
}

Other contributors:
\itemize{
  \item Robin Daelemans \email{robin.daelemans@inbo.be} (\href{https://orcid.org/0000-0003-0887-3140}{ORCID}) (Research Institute for Nature and Forest (INBO)) [contributor]
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) (Research Institute for Nature and Forest (INBO)) [contributor]
  \item Research Institute for Nature and Forest (INBO) (00j54wy13) [copyright holder]
  \item European Union (https://doi.org/10.3030/101112823) [funder]
}

}
\keyword{internal}
