% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_ESAS_Upload_Matrix.R
\name{Export_ESAS_Upload_Matrix}
\alias{Export_ESAS_Upload_Matrix}
\title{Write a tab delimited csv file as expected by the ESAS upload module.}
\usage{
Export_ESAS_Upload_Matrix(table, path, filename)
}
\arguments{
\item{table}{A data frame or matrix to be exported as returned by
\code{\link[=Transform_ESAS_Tables_4_Upload]{Transform_ESAS_Tables_4_Upload()}}}

\item{path}{Directory path where the file will be saved.}

\item{filename}{Character string specifying the name of the output file
without the extension.}
}
\value{
(invisible) The path to the exported file.
}
\description{
This function exports a data frame to a tab-delimited file without row names,
column names, or quotes, using the specified file encoding.
}
\examples{
#Read 4 ESAS tables:
data_path <- system.file("extdata", "ESAS_INBO_202401", package = "esas")
esas_tables_list <- Read_ESAS_Tables(path = data_path,
                                     file_encoding = "UTF-8")

#Convert tables to upload format:
esas_4_upload <- Transform_ESAS_Tables_4_Upload(campaigns_tbl = esas_tables_list$CAMPAIGNS,
                                                samples_tbl = esas_tables_list$SAMPLES,
                                                positions_tbl = esas_tables_list$POSITIONS,
                                                observations_tbl = esas_tables_list$OBSERVATIONS,
                                                data_provider = "202",
                                                country = "BE")

\dontrun{
#Export upload table:
Export_ESAS_Upload_Matrix(table = esas_4_upload,
                          path = tempdir(),
                          filename = "esas_4_upload")
}
}
\seealso{
Other upload functions: 
\code{\link{Transform_ESAS_Tables_4_Upload}()}
}
\concept{upload functions}
