% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_ESAS_Tables.R
\name{Read_ESAS_Tables}
\alias{Read_ESAS_Tables}
\title{Read csv files and combine them into a single data.frame}
\usage{
Read_ESAS_Tables(path, file_encoding = "UTF-8")
}
\arguments{
\item{path}{(Character.) Path to the directory containing 4 csv files as
returned by the \href{https://esas.ices.dk/inventory}{ESAS Data inventory}.
It should contain the files: "Campaigns.csv", "Samples.csv",
"Positions.csv" and "Observations.csv".}

\item{file_encoding}{(Character.) File encoding. Default is "UTF-8".}
}
\value{
A list of 4 data.frames: CAMPAIGNS, SAMPLES, POSITIONS, OBSERVATIONS.
}
\description{
Read 4 csv files (CAMPAIGNS, SAMPLES, POSITIONS, OBSERVATIONS) and combine
them into a list of data.frames.
}
\examples{
data_path <- system.file("extdata", "ESAS_0827343782", package = "esas")
Read_ESAS_Tables(data_path)

}
\concept{read data from ESAS}
