% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_ESAS_Table.R
\name{Create_ESAS_Table}
\alias{Create_ESAS_Table}
\title{Combine the tables from the ESAS Data Model into a single data.frame}
\usage{
Create_ESAS_Table(esas_tables_list)
}
\arguments{
\item{esas_tables_list}{A list containing the ESAS tables as data.frames. The
list should include the following named elements: CAMPAIGNS, SAMPLES,
POSITIONS, and OBSERVATIONS.}
}
\value{
A single data.frame that combines the information from the CAMPAIGNS,
SAMPLES, POSITIONS, and OBSERVATIONS tables.
}
\description{
The Analysis functions in this R package expect a single data.frame as input.
This function combines the tables from the ESAS Data Model into a single
data.frame. It performs left joins to ensure that all records from the
CAMPAIGNS, SAMPLES, POSITIONS, and OBSERVATIONS tables are included. It also
renames the Notes columns from the CAMPAIGNS and SAMPLES tables to avoid
naming conflicts.
}
\seealso{
Other analysis functions: 
\code{\link{Calculate_Detection_P_Ship_Based_Surveys}()},
\code{\link{Create_Seabird_Density_Cross_Table}()}
}
\concept{analysis functions}
