% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Seabird_Density_Cross_Table.R
\name{Create_Seabird_Density_Cross_Table}
\alias{Create_Seabird_Density_Cross_Table}
\title{Create a cross-table with distance-corrected bird densities}
\usage{
Create_Seabird_Density_Cross_Table(
  esas_table,
  probabilities,
  species_selection
)
}
\arguments{
\item{esas_table}{A data frame containing seabird observation data as
resulting from the \code{\link[=Create_ESAS_Table]{Create_ESAS_Table()}} function.}

\item{probabilities}{A data frame containing detection probabilities for
different species as returned by
\code{\link[=Calculate_Detection_P_Ship_Based_Surveys]{Calculate_Detection_P_Ship_Based_Surveys()}}}

\item{species_selection}{A vector of species codes to include in the analysis
as encoded in the species column of the Observations table in the ESAS Data
Model. See the \href{https://esas-docs.ices.dk/species/}{Species} page of the
Data Model.}
}
\value{
A data frame containing distance-corrected bird densities for the
selected species.
}
\description{
This function processes seabird observation data to create a cross-table of
distance-corrected bird densities for selected species. It filters the data
based on specific criteria, applies detection probability corrections, and
calculates densities per unit area. The resulting table includes relevant
metadata such as date, time, area, and location.
}
\examples{

path_to_read <- system.file("extdata", "ESAS_0827343782", package = "esas")
esas_table <- Create_ESAS_Table(Read_ESAS_Tables(path = path_to_read))
Create_Seabird_Density_Cross_Table(
  esas_table = esas_table,
  probabilities =
    Calculate_Detection_P_Ship_Based_Surveys(esas_table,
                                             species_2_analyse = c(720,6000)),
  species_selection = c(720, 6020)
)
}
\seealso{
Other analysis functions: 
\code{\link{Calculate_Detection_P_Ship_Based_Surveys}()},
\code{\link{Create_ESAS_Table}()}
}
\concept{analysis functions}
