% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_ESAS_Upload_Matrix.R
\name{Export_ESAS_Upload_Matrix}
\alias{Export_ESAS_Upload_Matrix}
\title{Write a tab delimited csv file as expected by the ESAS upload module.}
\usage{
Export_ESAS_Upload_Matrix(table, path, filename, file_encoding = "UTF-8")
}
\arguments{
\item{table}{A data frame or matrix to be exported as returned by
\code{\link[=Transform_ESAS_Tables_4_Upload]{Transform_ESAS_Tables_4_Upload()}}}

\item{path}{Directory path where the file will be saved.}

\item{filename}{Character string specifying the name of the output file
without the extension.}

\item{file_encoding}{Character string specifying the file encoding (default
is "UTF-8"). See \code{\link[base:connections]{base::file()}} for details on supported encodings.}
}
\value{
(invisible) The path to the exported file.
}
\description{
This function exports a data frame to a tab-delimited file without row names,
column names, or quotes, using the specified file encoding.
}
\seealso{
Other upload functions: 
\code{\link{Transform_ESAS_Tables_4_Upload}()}
}
\concept{upload functions}
